Move getLinkIconSvg to utils
